Refactors clients into one
