+++ src/backend/libs/efiling-cso-starter/src/main/java/ca/bc/gov/open/jag/efilingcsostarter/CsoHelpers.java
@@ -30,10 +30,8 @@ public class CsoHelpers {
      * @throws DatatypeConfigurationException
      */
     public static XMLGregorianCalendar date2XMLGregorian(Date date) throws DatatypeConfigurationException {
-
         GregorianCalendar c = new GregorianCalendar();
         c.setTime(date);
-        XMLGregorianCalendar date2 = DatatypeFactory.newInstance().newXMLGregorianCalendar(c);
-        return date2;
+        return DatatypeFactory.newInstance().newXMLGregorianCalendar(c);
     }
 }
