+++ src/backend/libs/efiling-cso-starter/src/main/java/ca/bc/gov/open/jag/efilingcsostarter/config/AutoConfiguration.java
@@ -8,7 +8,9 @@ import ca.bc.gov.open.jag.efilingcommons.model.EfilingSoapClientProperties;
 import ca.bc.gov.open.jag.efilingcommons.model.SoapProperties;
 import ca.bc.gov.open.jag.efilingcommons.service.EfilingAccountService;
 import ca.bc.gov.open.jag.efilingcommons.service.EfilingDocumentService;
+import ca.bc.gov.open.jag.efilingcommons.service.EfilingLookupService;
 import ca.bc.gov.open.jag.efilingcsostarter.CSODocumentServiceImpl;
+import ca.bc.gov.open.jag.efilingcsostarter.CsoLookupServiceImpl;
 import ca.bc.gov.open.jag.efilingcsostarter.CsoAccountServiceImpl;
 import ca.bc.gov.open.jag.efilingcsostarter.mappers.AccountDetailsMapper;
 import ca.bc.gov.open.jag.efilingcsostarter.mappers.AccountDetailsMapperImpl;
@@ -19,7 +21,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
+import ca.bc.gov.ag.csows.lookups.LookupFacadeBean;
 
 
 @Configuration
@@ -50,6 +52,9 @@ public class AutoConfiguration {
     public FilingStatusFacadeBean filingStatusFacadeBean() { return getPort(Clients.STATUS, FilingStatusFacadeBean.class); }
 
     @Bean
+    public LookupFacadeBean lookupFacadeBean() { return getPort(Clients.LOOKUP, LookupFacadeBean.class); }
+
+    @Bean
     public AccountDetailsMapper accountDetailsMapper() {
         return new AccountDetailsMapperImpl();
     }
@@ -68,6 +73,12 @@ public class AutoConfiguration {
     @ConditionalOnMissingBean({EfilingDocumentService.class})
     public EfilingDocumentService efilingDocumentService(FilingStatusFacadeBean filingStatusFacadeBean) {
         return new CSODocumentServiceImpl(filingStatusFacadeBean);
+    }
+
+    @Bean
+    @ConditionalOnMissingBean({EfilingLookupService.class})
+    public EfilingLookupService efilingLookupService (LookupFacadeBean lookupFacadeBean) {
+        return new CsoLookupServiceImpl(lookupFacadeBean);
     }
 
     public <T> T getPort(Clients clients, Class<T> type) {
