+++ src/backend/libs/efiling-cso-starter/src/test/java/ca/bc/gov/open/jag/efilingcsostarter/config/AutoConfigurationTest.java
@@ -4,6 +4,8 @@ package ca.bc.gov.open.jag.efilingcsostarter.config;
 import ca.bc.gov.open.jag.efilingcommons.model.Clients;
 import ca.bc.gov.open.jag.efilingcommons.model.EfilingSoapClientProperties;
 import ca.bc.gov.open.jag.efilingcommons.model.SoapProperties;
+import ca.bc.gov.open.jag.efilingcsostarter.CSODocumentServiceImpl;
+import ca.bc.gov.open.jag.efilingcsostarter.CsoLookupServiceImpl;
 import ca.bc.gov.open.jag.efilingcsostarter.CsoAccountServiceImpl;
 import ca.bc.gov.open.jag.efilingcsostarter.mappers.AccountDetailsMapperImpl;
 import org.junit.jupiter.api.*;
@@ -36,8 +38,12 @@ public class AutoConfigurationTest {
         Assertions.assertNotNull(sut.roleRegistryPortType());
         Assertions.assertNotNull(sut.bCeIDServiceSoap());
         Assertions.assertEquals(CsoAccountServiceImpl.class, sut.efilingAccountService(null, null, null, null).getClass());
+        Assertions.assertEquals(CSODocumentServiceImpl.class, sut.efilingDocumentService(null).getClass());
+        Assertions.assertEquals(CsoLookupServiceImpl.class, sut.efilingLookupService(null).getClass());
         Assertions.assertEquals(AccountDetailsMapperImpl.class, sut.accountDetailsMapper().getClass());
     }
+
+
 
     private SoapProperties initSoapProperties() {
