--- src/backend/libs/efiling-submission-client/src/main/java/ca/bc/gov/open/jag/efilingsubmissionclient/EfilingSubmissionService.java
@@ -1,11 +1,0 @@
-package ca.bc.gov.open.jag.efilingsubmissionclient;
-
-import ca.bc.gov.ag.csows.filing.FilingPackage;
-
-import java.math.BigDecimal;
-
-
-public interface EfilingSubmissionService {
-
-    public BigDecimal submitFiling(FilingPackage filingPackage);
-}
