--- src/backend/libs/efiling-submission-client/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  ca.bc.gov.open.jag.efilingsubmissionclient.config.AutoConfiguration
